# Utilize ArrayPool for ECC calculations