--- a/QRCoder/QRCodeGenerator.Polynom.cs
+++ b/QRCoder/QRCodeGenerator.Polynom.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using System.Threading;
 
 namespace QRCoder
 {