Improves performance by calculating `generatorPolynom` only a single time and renting arrays from `ArrayPool`.  Basically I rewrote `Polynom` to function similar to `List<PolynomItem>` with methods such as `Add`, `Clear` and `Sort`.  It's a struct so it does not require a heap allocation itself (no change), and the underlying buffer is rented from the array pool.  Calling `Dispose` on the `Polynom` will return the array back to the array pool.

## Before

| Method           | Mean       | Error   | StdDev  | Gen0    | Allocated |
|----------------- |-----------:|--------:|--------:|--------:|----------:|
| CreateQRCode     |   107.2 us | 0.66 us | 0.62 us |  1.2207 |  23.18 KB |
| CreateQRCodeLong | 1,878.4 us | 4.68 us | 4.15 us | 23.4375 | 433.53 KB |

## After - via ArrayPool<T>.Shared

| Method           | Mean        | Error    | StdDev   | Gen0   | Allocated |
|----------------- |------------:|---------:|---------:|-------:|----------:|
| CreateQRCode     |    99.44 us | 0.506 us | 0.448 us | 0.7324 |  14.68 KB |
| CreateQRCodeLong | 1,637.04 us | 2.811 us | 2.347 us | 1.9531 |  45.75 KB |

## After - via custom array pool implementation (for use with .NET Framework)

| Method           | Mean        | Error    | StdDev   | Gen0   | Allocated |
|----------------- |------------:|---------:|---------:|-------:|----------:|
| CreateQRCode     |    98.51 us | 0.188 us | 0.157 us | 0.7324 |  14.68 KB |
| CreateQRCodeLong | 1,706.09 us | 2.382 us | 2.112 us | 1.9531 |  45.75 KB |