@Shane32 2024-05-07
I don't know why these tests pass.  `CalculateECCWords` mutates `generatorPolynom`.  I have to rework this PR.