@Shane32 2024-05-07
Just for fun, I've added another benchmark for a 2600-character QR code and compared this PR against v1.5.1.  For these unreasonably-long QR codes, memory requirements have been reduced from 110 MB to 114 KB, a 99.9% reduction.  And there's still room for improvement.

## Version 1.5.1

| Method              | Mean        | Error     | StdDev    | Gen0      | Gen1     | Allocated    |
|-------------------- |------------:|----------:|----------:|----------:|---------:|-------------:|
| CreateQRCode        |    245.4 us |   2.36 us |   1.97 us |   17.0898 |        - |    317.99 KB |
| CreateQRCodeLong    |  7,492.1 us |  30.09 us |  26.67 us |  265.6250 |  15.6250 |   4920.13 KB |
| CreateQRCodeLongest | 88,024.3 us | 209.67 us | 185.86 us | 5833.3333 | 833.3333 | 110090.17 KB |

## This PR

| Method              | Mean         | Error     | StdDev    | Gen0   | Allocated | % RAM reduction |
|-------------------- |-------------:|----------:|----------:|-------:|----------:|----:|
| CreateQRCode        |     98.12 us |  0.213 us |  0.189 us | 0.7324 |  14.68 KB | 95% |
| CreateQRCodeLong    |  1,675.42 us |  1.812 us |  1.415 us | 1.9531 |  45.75 KB | 99% |
| CreateQRCodeLongest | 14,824.82 us | 48.627 us | 45.486 us |      - | 113.96 KB | 99.9% |