@Shane32 2024-05-09
Thanks @gfoidl !  I'll merge in your suggestions here later tonight.